Renaming test context configuration
