+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/BulkPrintServiceTest.java
@@ -25,7 +25,7 @@ import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-@ActiveProfiles("test-bulk-print-service")
+@ActiveProfiles("test-mock-document-client")
 public class BulkPrintServiceTest extends BaseServiceTest {
 
     @Autowired
